refactor(Board): Board 인터페이스 생성 및 ChessBoard 구현체 분리
