+++ src/main/java/chess/domain/ChessGame.java
@@ -3,6 +3,7 @@ package chess.domain;
 import chess.GameStatus;
 import chess.action.Action;
 import chess.domain.board.Board;
+import chess.domain.board.ChessBoard;
 import chess.domain.piece.Color;
 import chess.domain.position.Position;
 
@@ -17,7 +18,7 @@ public class ChessGame implements Action {
     
     public ChessGame() {
         this.status = GameStatus.READY;
-        this.board = Board.create();
+        this.board = ChessBoard.create();
         this.turn = Color.WHITE;
     }
