+++ src/test/java/chess/domain/ChessBoardTest.java
@@ -3,7 +3,7 @@ package chess.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.domain.board.Board;
+import chess.domain.board.ChessBoard;
 import chess.domain.piece.Color;
 import chess.domain.piece.Empty;
 import chess.domain.piece.Knight;
@@ -17,16 +17,16 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-class BoardTest {
+class ChessBoardTest {
     
     @Nested
     @DisplayName("생성 테스트")
-    class CreateBoard {
+    class CreateChessBoard {
         
         @Test
         @DisplayName("보드 피스 생성 테스트")
         void initialize2() {
-            Board board = Board.create();
+            ChessBoard board = ChessBoard.create();
             List<String> positions = List.of("a1", "b1", "g7", "h8", "d4");
             List<Piece> pieces = List.of(Rook.create(Color.WHITE), Knight.create(Color.WHITE), Pawn.create(Color.BLACK),
                     Rook.create(Color.BLACK), Empty.create());
@@ -44,12 +44,12 @@ class BoardTest {
         @Test
         @DisplayName("경로에 다른 피스가 있을 경우")
         void checkOtherPieceInRoute() {
-            Board board = Board.create();
+            ChessBoard board = ChessBoard.create();
             board.move(Position.from("a2"), Position.from("a4"));
             board.move(Position.from("a7"), Position.from("a5"));
             assertThatThrownBy(() -> board.checkRoute(Position.from("a1"), Position.from("a5"))).isInstanceOf(
                             IllegalArgumentException.class)
-                    .hasMessage(Board.OTHER_PIECE_IN_ROUTE);
+                    .hasMessage(ChessBoard.OTHER_PIECE_IN_ROUTE);
         }
     }
     
@@ -60,21 +60,21 @@ class BoardTest {
         @Test
         @DisplayName("위나 아래로 움직이는데, 경로에 다른 피스가 있을 경우")
         void checkOtherPieceInRoute() {
-            Board board = Board.create();
+            ChessBoard board = ChessBoard.create();
             board.move(Position.from("a2"), Position.from("a4"));
             board.move(Position.from("a7"), Position.from("a5"));
             assertThatThrownBy(() -> board.checkRoute(Position.from("a4"), Position.from("a5"))).isInstanceOf(
                             IllegalArgumentException.class)
-                    .hasMessage(Board.OTHER_PIECE_IN_ROUTE);
+                    .hasMessage(ChessBoard.OTHER_PIECE_IN_ROUTE);
         }
         
         @Test
         @DisplayName("대각선 방향으로 움직이는데, 상대편 피스가 없는 경우 - 비어있는 경우")
         void checkOtherPieceInDiagonal1() {
-            Board board = Board.create();
+            ChessBoard board = ChessBoard.create();
             assertThatThrownBy(() -> board.checkRoute(Position.from("a2"), Position.from("b3"))).isInstanceOf(
                             IllegalArgumentException.class)
-                    .hasMessage(Board.NO_OTHER_COLOR_IN_DIAGONAL_DESTINATION);
+                    .hasMessage(ChessBoard.NO_OTHER_COLOR_IN_DIAGONAL_DESTINATION);
         }
         
     }
@@ -82,20 +82,20 @@ class BoardTest {
     @Test
     @DisplayName("움직이는 대상이 다른 색깔인지 확인")
     void checkColor() {
-        Board board = Board.create();
+        ChessBoard board = ChessBoard.create();
         assertThatThrownBy(() -> board.checkColor(Position.from("a2"), Position.from("a4"), Color.BLACK)).isInstanceOf(
                         IllegalArgumentException.class)
-                .hasMessage(Board.OTHER_COLOR_IN_SOURCE);
+                .hasMessage(ChessBoard.OTHER_COLOR_IN_SOURCE);
     }
     
     @Test
     @DisplayName("목적지에 같은 색깔의 피스가 있는지 확인")
     void checkDestination() {
-        Board board = Board.create();
+        ChessBoard board = ChessBoard.create();
         board.move(Position.from("b2"), Position.from("a4"));
         assertThatThrownBy(
                 () -> board.checkColor(Position.from("a2"), Position.from("a4"), Color.WHITE)).isInstanceOf(
                         IllegalArgumentException.class)
-                .hasMessage(Board.SAME_COLOR_IN_DESTINATION);
+                .hasMessage(ChessBoard.SAME_COLOR_IN_DESTINATION);
     }
 }
